Turn update server autosave function into a redux thunk action
