+++ package.json
@@ -77,6 +77,7 @@
     "redux-logger": "3.0.6",
     "redux-thunk": "2.3.0",
     "stackframe": "1.0.4",
+    "throttle-action": "0.1.0",
     "universal-router": "8.1.0",
     "webpack-shell-plugin": "0.5.0"
   },
